ci: update Jenkins file for signing
Change durabilityHint to MAX_SURVIVABILITY due to a recurring error:
ERROR: Cannot resume build because FlowNode could not be loaded. This is expected to happen when using the PERFORMANCE_OPTIMIZED durability setting and Jenkins is not shut down cleanly. Consider investigating to understand if Jenkins was not shut down cleanly or switching to the MAX_SURVIVABILITY durability setting which should prevent this issue in most cases.