Improve update workflow
Rename `updates.reportOnStart` to  `updates.checkForUpdates`, because it no longer refers to only the startup.
Introduce `updates.checkInterval` which is the interval of update checks in minutes, when `updates.checkForUpdates` is enabled.
Default for `updates.checkInterval` is 60 minutes.
Also changed it so if the user disables the `updates.checkForUpdates` no request is sent to the download page.

Closes #546.

Signed-off-by: Simon Graband <sgraband@eclipsesource.com>